Better organization of functions.
